cancer/requirements.txt
# Core dependencies for data science and ML
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
jupyter>=1.0.0
ipykernel>=6.0.0

# Deep Learning
tensorflow>=2.10.0
keras>=2.10.0

# Computer Vision and Medical Imaging
opencv-python>=4.6.0
Pillow>=9.0.0
scikit-image>=0.19.0
pydicom>=2.3.0
SimpleITK>=2.2.0

# Machine Learning
scikit-learn>=1.1.0
scipy>=1.9.0

# Radiomics (optional - may need manual installation)
# pyradiomics>=3.0.1

# Web Dashboard
streamlit>=1.25.0
streamlit-option-menu>=0.3.0

# Google AI
google-generativeai>=0.3.0

# API and Web Requests
requests>=2.28.0
urllib3>=1.26.0

# Data Processing
tqdm>=4.64.0
joblib>=1.2.0

# Configuration and Utils
python-dotenv>=0.19.0
pydantic>=2.7.0
pathlib2>=2.3.0
configparser>=5.3.0

# Visualization and Reporting
jupyterlab>=3.4.0
ipywidgets>=8.0.0

# Medical Data Formats
nibabel>=4.0.0  # For NIfTI files

# Statistical Analysis
statsmodels>=0.13.0
pingouin>=0.5.0  # Advanced statistics

# Image Enhancement
imageio>=2.22.0
# skimage alias eliminado; usar scikit-image

# Development and Testing (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0